+++ tiger-common/src/main/java/de/gematik/test/tiger/common/config/PkiType.java
@@ -1,0 +1,7 @@
+package de.gematik.test.tiger.common.config;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+
+public enum PkiType {
+    Certificate, Key
+}