+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/TestTigerProxyHeavyDuty.java
@@ -69,7 +69,7 @@ public class TestTigerProxyHeavyDuty {
                 .to("http://localhost:" + wireMockRule.port())
                 .build()))
             .proxyLogLevel("WARN")
-            .disableRbelParsing(true)
+            .activateRbelParsing(false)
             .build());
 
         final UnirestInstance unirestInstance = Unirest.spawnInstance();